Board Limit and User Authentication verify

+++ kanbanapi/src/main/java/sit/int221/kanbanapi/services/BoardService.java
@@ -7,7 +7,9 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import sit.int221.kanbanapi.databases.kanbandb.entities.Board;
 import sit.int221.kanbanapi.databases.kanbandb.repositories.BoardRepository;
+import sit.int221.kanbanapi.databases.userdb.entities.User;
 import sit.int221.kanbanapi.databases.userdb.repositories.UserRepository;
+import sit.int221.kanbanapi.exceptions.AuthenticationFailed;
 import sit.int221.kanbanapi.exceptions.BadRequestException;
 
 import java.util.List;
@@ -18,7 +20,7 @@ public class BoardService {
     private BoardRepository boardRepository;
 
     @Autowired
-    UserRepository userRepository;
+    private UserRepository userRepository;
 
     public List<Board> getUserBoards(String user) {
         return boardRepository.findByOwnerId(userRepository.findByUsername(user).getOid());
@@ -38,20 +40,35 @@ public class BoardService {
         board.setBoardId(boardId);
         board.setBoardName(boardName);
         board.setOwnerId(userRepository.findByUsername(owner.getUsername()).getOid());
+        board.setTaskLimitEnabled(Boolean.FALSE);
+        board.setMaxTasksPerStatus(10);
         return boardRepository.save(board);
     }
 
     @Transactional
-    public Board removeBoard(String id) {
-        Board board = boardRepository.findById(id).orElseThrow(() -> new BadRequestException("Board "+ id + " does not exist"));
+    public Board removeBoard(String boardId, String username) {
+        checkBoardOwnership(boardId, username);
+        Board board = boardRepository.findById(boardId).orElseThrow(() -> new BadRequestException("Board "+ boardId + " does not exist"));
         boardRepository.delete(board);
         return board;
     }
 
     @Transactional
-    public Board updateBoard(String id, String boardName) {
-        Board board = getBoardById(id);
-        board.setBoardName(boardName);
+    public Board updateBoard(String boardId, Board newBoard, String username) {
+        checkBoardOwnership(boardId, username);
+        Board board = getBoardById(boardId);
+        board.setBoardName(newBoard.getBoardName());
+        board.setTaskLimitEnabled(newBoard.getTaskLimitEnabled());
+        board.setMaxTasksPerStatus(newBoard.getMaxTasksPerStatus());
         return boardRepository.save(board);
+    }
+
+    public User checkBoardOwnership(String boardId, String username) {
+        Board board = getBoardById(boardId);
+        User owner = userRepository.findByUsername(board.getOwnerId());
+        if (!owner.getUsername().equals(username)) {
+            throw new AuthenticationFailed("You do not have permission to perform this action.");
+        }
+        return owner;
     }
 }
